add dataclass config
