--- src/test_run.py
@@ -1,6 +1,0 @@
-import time
-import os
-
-
-print("listing dir:")
-print(os.listdir("/app/data"))
